feat: consolidate data and privacy and about pages into settings page
